+++ lib/Markdown/src/MarkdownParser.h
@@ -38,7 +38,7 @@ class MarkdownParser : public ContentParser {
   MarkdownParser(std::string filepath, GfxRenderer& renderer, const RenderConfig& config);
   ~MarkdownParser() override;
 
-  bool parsePages(const std::function<void(std::unique_ptr<Page>)>& onPageComplete, uint16_t maxPages = 0,
+  bool parsePages(const std::function<void(std::unique_ptr<Page>)>& onPageComplete, uint32_t maxPages = 0,
                   const AbortCallback& shouldAbort = nullptr) override;
   bool hasMoreContent() const override { return hasMore_; }
   bool canResume() const override { return currentOffset_ > 0 && hasMore_; }
@@ -83,8 +83,8 @@ class MarkdownParser : public ContentParser {
     bool inCodeBlock;
     int headerLevel;
     bool hitMaxPages;
-    uint16_t pagesCreated;
-    uint16_t maxPages;
+    uint32_t pagesCreated;
+    uint32_t maxPages;
     std::function<void(std::unique_ptr<Page>)> onPageComplete;
     AbortCallback shouldAbort;
