+++ lib/PageCache/src/HtmlParser.h
@@ -24,11 +24,11 @@ class HtmlParser : public ContentParser {
   bool initialized_ = false;
 
   std::function<void(std::unique_ptr<Page>)> onPageComplete_;
-  uint16_t maxPages_ = 0;
-  uint16_t pagesCreated_ = 0;
+  uint32_t maxPages_ = 0;
+  uint32_t pagesCreated_ = 0;
   bool hitMaxPages_ = false;
 
-  std::vector<std::pair<std::string, uint16_t>> anchorMap_;
+  std::vector<std::pair<std::string, uint32_t>> anchorMap_;
 
   void cleanupTempFiles();
 
@@ -36,12 +36,12 @@ class HtmlParser : public ContentParser {
   HtmlParser(std::string filepath, std::string cacheDir, GfxRenderer& renderer, const RenderConfig& config);
   ~HtmlParser() override;
 
-  bool parsePages(const std::function<void(std::unique_ptr<Page>)>& onPageComplete, uint16_t maxPages = 0,
+  bool parsePages(const std::function<void(std::unique_ptr<Page>)>& onPageComplete, uint32_t maxPages = 0,
                   const AbortCallback& shouldAbort = nullptr) override;
   bool hasMoreContent() const override { return hasMore_; }
   bool canResume() const override { return initialized_ && liveParser_ != nullptr; }
   void reset() override;
-  const std::vector<std::pair<std::string, uint16_t>>& getAnchorMap() const override;
+  const std::vector<std::pair<std::string, uint32_t>>& getAnchorMap() const override;
   uint32_t bytesConsumed() const override;
   uint32_t totalBytes() const override;
 };
