+++ lib/PageCache/src/PlainTextParser.h
@@ -43,7 +43,7 @@ class PlainTextParser : public ContentParser {
   PlainTextParser(std::string filepath, GfxRenderer& renderer, const RenderConfig& config);
   ~PlainTextParser() override = default;
 
-  bool parsePages(const std::function<void(std::unique_ptr<Page>)>& onPageComplete, uint16_t maxPages = 0,
+  bool parsePages(const std::function<void(std::unique_ptr<Page>)>& onPageComplete, uint32_t maxPages = 0,
                   const AbortCallback& shouldAbort = nullptr) override;
   bool hasMoreContent() const override { return hasMore_; }
   bool canResume() const override { return currentOffset_ > 0 && hasMore_; }
