+++ src/content/GlobalPageMetrics.h
@@ -9,7 +9,7 @@ namespace papyrix::page_metrics {
 constexpr size_t kEstimatedBytesPerPage = 2048;
 
 struct Section {
-  uint16_t pages = 0;
+  uint32_t pages = 0;
   bool exact = false;
   uint32_t byteSize = 0;
 };
@@ -25,8 +25,8 @@ struct Update {
   bool becameExact = false;
 };
 
-uint16_t estimatePagesForBytes(size_t bytes, size_t bytesPerPage = kEstimatedBytesPerPage);
-Update applyCache(Section& metric, uint16_t pageCount, bool partial);
+uint32_t estimatePagesForBytes(size_t bytes, size_t bytesPerPage = kEstimatedBytesPerPage);
+Update applyCache(Section& metric, uint32_t pageCount, bool partial);
 void fillEstimates(std::vector<Section>& metrics, size_t bytesPerPage = kEstimatedBytesPerPage);
 void recalibrate(std::vector<Section>& metrics);
 uint32_t total(const std::vector<Section>& metrics);
